+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 5.1.0-dev
 - added assertion assertNoException
+- Job reports total elapsed time in milliseconds in $totalTime
 
 Version 5.0.0
 - BC break: forbidden passing string to parameter $value of assertions assertCount and assertNotCount